chore: migrate datastore module into the gcloud module
