+++ .github/workflows/ci.yml
@@ -123,7 +123,7 @@ jobs:
     needs: test-modules
     strategy:
       matrix:
-        module: [cockroachdb, consul, datastore, firestore, nginx, pubsub, spanner, toxiproxy]
+        module: [cockroachdb, consul, firestore, nginx, pubsub, spanner, toxiproxy]
     uses: ./.github/workflows/ci-test-go.yml
     with:
       go-version: "1.20.x"
