--- docs/examples/datastore.md
@@ -1,9 +1,0 @@
-# Datastore
-
-<!--codeinclude-->
-[Creating a Datastore container](../../examples/datastore/datastore.go)
-<!--/codeinclude-->
-
-<!--codeinclude-->
-[Test for a Datastore container](../../examples/datastore/datastore_test.go)
-<!--/codeinclude-->
