+++ modules/gcloud/examples_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"cloud.google.com/go/bigtable"
+	"cloud.google.com/go/datastore"
 	"google.golang.org/api/option"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
@@ -17,14 +18,14 @@ func ExampleRunBigTableContainer() {
 	// runBigTableContainer {
 	ctx := context.Background()
 
-	gcloudContainer, err := gcloud.RunBigTableContainer(ctx, testcontainers.WithImage("gcr.io/google.com/cloudsdktool/cloud-sdk:367.0.0-emulators"))
+	bigTableContainer, err := gcloud.RunBigTableContainer(ctx, testcontainers.WithImage("gcr.io/google.com/cloudsdktool/cloud-sdk:367.0.0-emulators"))
 	if err != nil {
 		panic(err)
 	}
 
 	// Clean up the container
 	defer func() {
-		if err := gcloudContainer.Terminate(ctx); err != nil {
+		if err := bigTableContainer.Terminate(ctx); err != nil {
 			panic(err)
 		}
 	}()
@@ -37,7 +38,7 @@ func ExampleRunBigTableContainer() {
 	)
 
 	options := []option.ClientOption{
-		option.WithEndpoint(gcloudContainer.URI),
+		option.WithEndpoint(bigTableContainer.URI),
 		option.WithoutAuthentication(),
 		option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
 	}
@@ -76,4 +77,57 @@ func ExampleRunBigTableContainer() {
 
 	// Output:
 	// Gopher
+}
+
+func ExampleRunDatastoreContainer() {
+	// runDatastoreContainer {
+	ctx := context.Background()
+
+	datastoreContainer, err := gcloud.RunDatastoreContainer(ctx, testcontainers.WithImage("gcr.io/google.com/cloudsdktool/cloud-sdk:367.0.0-emulators"))
+	if err != nil {
+		panic(err)
+	}
+
+	// Clean up the container
+	defer func() {
+		if err := datastoreContainer.Terminate(ctx); err != nil {
+			panic(err)
+		}
+	}()
+	// }
+
+	options := []option.ClientOption{
+		option.WithEndpoint(datastoreContainer.URI),
+		option.WithoutAuthentication(),
+		option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
+	}
+	dsClient, err := datastore.NewClient(ctx, "test-project", options...)
+	if err != nil {
+		panic(err)
+	}
+	defer dsClient.Close()
+
+	type Task struct {
+		Description string
+	}
+
+	k := datastore.NameKey("Task", "sample", nil)
+	data := Task{
+		Description: "my description",
+	}
+	_, err = dsClient.Put(ctx, k, &data)
+	if err != nil {
+		panic(err)
+	}
+
+	saved := Task{}
+	err = dsClient.Get(ctx, k, &saved)
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Println(saved.Description)
+
+	// Output:
+	// my description
 }
