+++ modules/gcloud/go.mod
@@ -4,6 +4,7 @@ go 1.20
 
 require (
 	cloud.google.com/go/bigtable v1.19.0
+	cloud.google.com/go/datastore v1.13.0
 	github.com/testcontainers/testcontainers-go v0.24.1
 	google.golang.org/api v0.142.0
 	google.golang.org/grpc v1.58.1
